Move protocol, utils, uv into src/server/ module
Create src/server/ as a peer to src/native/. Move protocol/, utils.rs,
uv.rs, and uv-install.sh into it. Update lib.rs to use a single
cfg-gated `mod server` instead of three separate gated modules.
