+++ src/server/mod.rs
@@ -1,0 +1,8 @@
+// Server backend for Hegel.
+//
+// When the `native` feature is NOT enabled, this module provides the test
+// runner that communicates with a Python hegel-core server over Unix sockets.
+
+pub(crate) mod protocol;
+pub(crate) mod utils;
+pub(crate) mod uv;
