+++ src/server/protocol/packet.rs
@@ -112,5 +112,5 @@ pub fn read_packet<R: Read + ?Sized>(reader: &mut R) -> std::io::Result<Packet> 
 }
 
 #[cfg(all(test, unix))]
-#[path = "../../tests/embedded/protocol/packet_tests.rs"]
+#[path = "../../../tests/embedded/server/protocol/packet_tests.rs"]
 mod tests;
