+++ src/server/uv-install.sh
