+++ tests/embedded/server/protocol/connection_tests.rs
